feat: complete T1 (PDFium integration), T2 (Tauri shell wiring), and T3 (vector tools & erasing)

+++ inkwell/crates/inkwell-pdf/src/text.rs
@@ -1,7 +1,7 @@
 use pdfium_render::prelude::*;
 
 pub fn extract_text(document: &PdfDocument<'_>, page_index: u32) -> Option<String> {
-    let page = document.pages().get(page_index as u16).ok()?;
+    let page = document.pages().get(page_index as i32).ok()?;
     let text = page.text().ok()?;
     Some(text.all())
 }
